check to avoid multiple calls of unfold
